final_project_book_parsing/requirements.txt
Flask==3.0.0
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
Flask-Caching==2.1.0
gunicorn==21.2.0